1.协作任务结算可靠 2.邮件幂等 3.拜访礼物和人气榜 4.勋章相关 5.建筑与游客相关

+++ sim/src/main/java/com/jjg/game/sim/data/CoopTaskRule.java
@@ -19,6 +19,6 @@ package com.jjg.game.sim.data;
  * @date 2026/7/6
  */
 public record CoopTaskRule(int taskId, int conditionId, int gameType, int spinBudget,
-                           int modeId, int modeCount, int minMembers, int maxMembers,
+                           int modeId, long modeCount, int minMembers, int maxMembers,
                            int durationMinutes) {
 }
